requirements.txt
flask
flask-cors
Pillow
numpy
gradio==3.50.2
datadog_api_client